### 2020-03-04 #2140 @whedon
[PRE REVIEW]: Jabberwocky: an ontology-aware toolkit for manipulating text
**Submitting author:** @sap218 (<a href="http://orcid.org/0000-0002-6169-0135">Samantha Pendleton</a>)
**Repository:** <a href="https://github.com/sap218/jabberwocky" target ="_blank">https://github.com/sap218/jabberwocky</a>
**Version:** v0.5.0.1
