### 2020-03-04 #2140 @Kevin-Mattheus-Moerman
@sap218 thanks for submitting this work. Can appears to be a re-submission of this work which was previously rejected by JOSS: https://github.com/openjournals/joss-reviews/issues/2054. Can you tell us in detail how the functionality of project has been significantly expanded with respect to the previously submitted version?
