### 2020-03-06 #2140 @sap218
@Kevin-Mattheus-Moerman Thanks!

Previously I submitted Jabberwocky with a single command, which allowed users to retrieve key elements from a text corpus using an ontology’s terms. Jabberwocky is the first tool to use an ontology directly. However, I understand the feedback that Jabberwocky was a minor-utility package and so I significantly expanded to include more functionality that demonstrates its utility in developing and maintaining ontologies, as well as querying text. This updated version of Jabberwocky includes two new subcommands, which together combine to make a workflow for users. Jabberwocky can now perform a tf-idf analysis on textual data to select key terms, and annotate those key terms in the ontology. Users can iteratively use these commands to identify and incrementally add new synonyms to expand their ontology.
