### 2020-03-15 #2140 @majensen
@wdduncan -- would you be willing to review [this software]() for [JOSS](https://joss.theoj.org)? Information about reviewing is [available here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Please let me know here or via other channels - thanks for considering! --@majensen
