### 2020-03-16 #2140 @majensen
@cmungall - would you have time and inclination to review [this software]() for [JOSS](https://joss.theoj.org)? Reviewing information is [available here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Let me know here or at maj.fortinbras@gmail.com if you have any questions. Thanks for considering! --@majensen
