### 2020-03-17 #2140 @majensen
To all - JOSS is operating under a [very relaxed schedule](https://github.com/openjournals/joss-reviews/issues/2161#issuecomment-598980193) during these days of COVID-19. @sap218, @wdduncan , please just drop a note here if you need to pause the review process to attend to more pressing matters.
