### 2020-03-19 #2168 @arfon
@sap218 - please change them to keys like this:

```
doi  = "10.1093/database/bau033"
```

i.e. change the `url` to `doi` and strip off the `https://doi.org/` preamble.
