+++ paper.bib
@@ -7,11 +7,10 @@
   journal  = "Database",
   volume   =  2014,
   number   =  0,
-  pages    = "bau033",
   month    =  apr,
   year     =  2014,
   language = "en",
-  url      = {https://doi.org/10.1093/database/bau033}
+  doi      = {10.1093/database/bau033}
 }
 
 @ARTICLE{Honnibal2017-dn,
@@ -32,7 +31,7 @@
                Finkel, Jenny and Bethard, Steven and McClosky, David",
   pages     = "55--60",
   year      =  2014,
-  url       = {https://doi.org/10.3115/v1/p14-5010}
+  doi       = {10.3115/v1/p14-5010}
 }
 
 @ARTICLE{Schriml2012-qp,
@@ -43,11 +42,10 @@
   journal  = "Nucleic Acids Res.",
   volume   =  40,
   number   = "Database issue",
-  pages    = "D940--6",
   month    =  jan,
   year     =  2012,
   language = "en",
-  url      = {https://doi.org/10.1093/nar/gkr972}
+  doi      = {10.1093/nar/gkr972}
 }
 
 @ARTICLE{Robinson2008-jh,
@@ -62,7 +60,7 @@
   month    =  nov,
   year     =  2008,
   language = "en",
-  url      = {https://doi.org/10.1016/j.ajhg.2008.09.017}
+  doi      = {10.1016/j.ajhg.2008.09.017}
 }
 
 @ARTICLE{Hoehndorf2015-qr,
@@ -77,7 +75,7 @@
   year     =  2015,
   keywords = "Semantic Web; data integration; data mining; ontology",
   language = "en",
-  url      = {https://doi.org/10.1093/bib/bbv011}
+  doi      = {10.1093/bib/bbv011}
 }
 
 
