### 2020-03-21 #2168 @sap218
Thanks @arfon - @majensen I believe I have fixed the DOI issue now
