### 2020-03-30 #2168 @wdduncan
I finished the checklist.

Comments:

1. I could not install the repo as per the instructions. When running the install script, I received error:
```
ERROR: TEST FAILED: /Users/wdduncan/.local/lib/python3.7/site-packages/ does NOT support .pth files
error: bad install directory or PYTHONPATH
```
This could be b/c I am using OS X, or perhaps b/c I installed using venv.

2.. The examples reference a file named `/ontology/pocketmonsters.owl`. But this file is not in the repo.

The github documentation would benefit from having examples of expected output and descriptions of what the output means.

The paper refers to the jabberwocky-tests as a way to see the software at work. But that directory contains no instructions.

How long is the software supposed paper to be?
