+++ paper.md
@@ -30,7 +30,7 @@ We introduce Jabberwocky, a Python-based [@Van_Rossum1995-ia], open-source toolk
 
 Jabberwocky also helps ontology developers to iteratively improve their ontology. The ``bite`` command allows a user to provide textual data and rank the important terms: using the term frequency–inverse document frequency (tf-idf) method from “scikit-learn” [@Pedregosa2011-st], which calculates an importance metric for a term based on the frequency of its occurrence and the document size. Providing an ontology will exclude terms already described in the ontology, meaning the result of ``bite`` will be a CSV of candidate terms to potentially be added to the ontology, exported by “pandas” [@McKinney2010-xf]. Once an expert has reviewed the terms and associated them to a class in the ontology, Jabberwocky’s third command ``arise`` will annotate the classes in the ontology, adding the newly identified synonyms. Iteratively performing multiple rounds of ``bite`` and ``arise`` can help the development and maintenance of ontologies. A user could use the ``catch`` command to confirm the modified ontology now captures more of the corpus.
 
-In Jabberwocky’s test repository (https://github.com/sap218/jabberwocky-tests), I show examples of each command separately. In the ‘process’ directory, I combine all three commands to demonstrate an example workflow. With 24 blog posts, the first use of ``catch`` returned 11 posts with the provided keywords. With ``bite`` I reviewed the CSV of ranked terms and curated new synonyms, simply by adding the corresponding class label from the ontology. I then used ``arise`` to add the identified synonyms into the ontology. With the second round of ``catch`` the number of posts returned for the same keywords increased to 16. This is a basic and straightforward example, but powerful. With Jabberwocky, users can efficiently search their text and gain more instances, providing new insight.
+In Jabberwocky’s test repository (see Jabberwocky repo for further instructions), I show examples of each command separately. In the ‘process’ directory, I combine all three commands to demonstrate an example workflow. With 24 blog posts, the first use of ``catch`` returned 11 posts with the provided keywords. With ``bite`` I reviewed the CSV of ranked terms and curated new synonyms, simply by adding the corresponding class label from the ontology. I then used ``arise`` to add the identified synonyms into the ontology. With the second round of ``catch`` the number of posts returned for the same keywords increased to 16. This is a basic and straightforward example, but powerful. With Jabberwocky, users can efficiently search their text and gain more instances, providing new insight.
 
 Jabberwocky leverages the strength of ontologies and text for a wide range of tasks. It will be useful to users who want to manipulate textual data using controlled vocabulary from ontologies.
 
