+++ README.md
@@ -2,7 +2,9 @@
 
 A toolkit for **ontologies**. Since we all know ontologies are "*nonsense*". **Note**: to check version, see `setup.py` in your local copy.
 
-### Elements
+Don't hesitate to create an [`issue`](https://github.com/sap218/jabberwocky/issues) or [`pull request`](https://github.com/sap218/jabberwocky/pulls) (see [**guidelines**](https://github.com/sap218/jabberwocky/blob/master/CONTRIBUTING.md) first).
+
+#### Elements
 
 command | description
 ------- | -----------
@@ -23,8 +25,8 @@ $ cd jabberwocky
 $ python3 setup.py install --user
 ```
 
-##### Examples
-Please see the [`jabberwocky-tests`](https://github.com/sap218/jabberwocky-tests)
+#### Examples
+For examples of Jabberwocky's commands in use, please see the **[`jabberwocky-tests`](https://github.com/sap218/jabberwocky-tests)** repository.
 
 ---
 
@@ -114,8 +116,6 @@ $ arise -o pato.owl -f new_synonyms.csv
 ## Thanks! :dragon:
 
 The poem "Jabberwocky" written by Lewis Carrol is described as a "nonsense" poem.
-
-Don't hesitate to create an [`issue`](https://github.com/sap218/jabberwocky/issues) or [`pull request`](https://github.com/sap218/jabberwocky/pulls) (see [**guidelines**](https://github.com/sap218/jabberwocky/blob/master/CONTRIBUTING.md) first).
 
 You can combine these commands together to form a process of steps of ontology synonym development and text analysis. See `jabberwocky-tests` repo for the [`jabberwocky-tests/process`](https://github.com/sap218/jabberwocky-tests/tree/master/process) directory for a chain of commands (as described in the image below).
 
