### 2020-04-17 sap218/jabberwocky#10 @balhoff
missing dependency?
I installed jabberwocky and tried to run the first example from jabberwocky-tests. I got this output:

```
(pyenv) localhost:catch jim (master)$ catch --ontology ../ontology/pocketmonsters.owl --keywords listofwords.txt --textfile blogs_formatted.json --parameter blog_post > catch_output.txt
Traceback (most recent call last):
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/bin/catch", line 11, in <module>
    load_entry_point('jabberwocky==0.5.0.1', 'console_scripts', 'catch')()
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/click/core.py", line 829, in __call__
    return self.main(*args, **kwargs)
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/click/core.py", line 782, in main
    rv = self.invoke(ctx)
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/click/core.py", line 1066, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/click/core.py", line 610, in invoke
    return callback(*args, **kwargs)
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/jabberwocky-0.5.0.1-py3.7.egg/catch/catch.py", line 426, in main
    search_term_dictionary = ontologyPurl(ontology, keywords)
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/jabberwocky-0.5.0.1-py3.7.egg/catch/catch.py", line 77, in ontologyPurl
    souped = souping(ontology) # Reading in the ontology file
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/jabberwocky-0.5.0.1-py3.7.egg/catch/catch.py", line 25, in souping
    soup = BeautifulSoup(contents,'xml')
  File "/Users/jim/Documents/Reviews/JOSS/pyenv/lib/python3.7/site-packages/bs4/__init__.py", line 228, in __init__
    % ",".join(features))
bs4.FeatureNotFound: Couldn't find a tree builder with the features you requested: xml. Do you need to install a parser library?
```

I'm running in a virtual environment with only the jabberwocky dependencies installed.

(this is part of JOSS review https://github.com/openjournals/joss-reviews/issues/2168)
