+++ requirements.txt
@@ -1,0 +1,4 @@
+click
+BeautifulSoup4
+scikit-learn
+pandas
