### 2020-05-01 sap218/jabberwocky#10 @sap218
it may be an issue with the 'jabberwocky-tests' repo documentation but the idea is that
`$ catch --ontology ../ontology/pocketmonsters.owl --keywords .............`
your -ontology file should direct where you clone the directory, e.g.
`--ontology /home/samantha/git-clones/jabberwocky-tests/ontology/pocketmonsters.owl`

and so on for the other files: the repo has the files, and you direct to them in the CLI:

`$ catch --ontology /home/samantha/git-clones/jabberwocky-tests/ontology/pocketmonsters.owl --keywords /home/samantha/git-clones/jabberwocky-tests/ontology/listofwords.txt --textfile /home/samantha/git-clones/jabberwocky-tests/ontology/blogs_formatted.json --parameter blog_post > catch_output.txt`
