### 2020-05-02 sap218/jabberwocky#10 @balhoff
Thanks @sap218—I am running in the 'jabberwocky-tests/catch' directory. The command line does seem to point to the right folders. If the paths are wrong, I get a different error about 'No such file', which makes sense.

I installed jabberwocky by first making a virtual environment: `python3 -m venv pyenv; source ./pyenv/bin/activate`. Then I installed all the packages using that python environment.
