+++ requirements.txt
@@ -2,3 +2,4 @@ click
 BeautifulSoup4
 scikit-learn
 pandas
+lxml
