### 2020-05-04 sap218/jabberwocky#10 @sap218
Thanks @balhoff - really sorry about that! It seems I did forget about a dependency! I've just updated the requirements.txt and README. you may need to re-clone.

I'm going to close this issue, but please reopen an issue if you find another issue!
