### 2020-05-16 #2168 @wdduncan
@sap218
I tried installing the software, but I am having issues. Steps so far:
1. Create venv using `python3 -m venv .env`, and active environment (works fine).
2. Run `pip install -r requirements.txt` (works fine).

But, when I run `python3 setup.py install --user`, I get message/error:
```
You are attempting to install a package to a directory that is not
on PYTHONPATH and which Python does not read ".pth" files from.  The
installation directory you specified (via --install-dir, --prefix, or
the distutils default setting) was:

    /Users/wdduncan/.local/lib/python3.7/site-packages/

and your PYTHONPATH environment variable currently contains:

    ''
```
I fixed this by removing `--user`. Perhaps it is related to my using a virtual environment? if so, it would be good to document this.

I ran the `jabberwocky-tests`.

For the `catch` test, I am unsure what the output in `catch_output.txt` means. Did `catch` find the lines in the blog that contained terms from the ontology?
Some documentation on what the output means would be helpful.

The `arise` test ran successfully. But, like the `catch` test, documentation would be helpful for describing the output.

The `arise` test:
```
arise -ontology ../ontology/pocketmonsters.owl -tfidf new_synonyms_tfidf.csv $ arise -o ../ontology/pocketmonsters.owl -f new_synonyms_tfidf.csv
```
 failed with error:
```
Usage: arise [OPTIONS]
Try 'arise --help' for help.

Error: no such option: -t
```
Hope this helps!

cc @majensen
