### 2020-05-16 sap218/jabberwocky#11 @sap218
create documentation for the jabberwocky-tests repo 
in relation to https://github.com/openjournals/joss-reviews/issues/2168
