### 2020-05-18 #2168 @wdduncan
@majensen I'm not sure, I usually don't use the `--user` option with `pip`, unless I run into permission issues. Since, I was testing the software, I didn't want to affect the library I have. My guess is that in the venv, there isn't a "user" specified. So, `--user` doesn't make sense.  But, I may be wrong about this ... I'm not a virtual environment guru :).

In any case, it is probably best document the behavior.
