### 2020-05-18 #2168 @majensen
Ah! Well I believe the virtual environment is truly isolated from both your system install and your user packages. Everything is stored in the venv, including the python binary, pip and everything else. But - better safe than sorry.
With python 3.7.6 (in an active venv) I get a more straightforward message:
`jensenma$ source venv/bin/activate
(venv) jensenma$ pip install --user requests
ERROR: Can not perform a '--user' install. User site-packages are not visible in this virtualenv.`

So I think it's a python thing, not a jabberwocky thing.
