### 2020-05-18 #2168 @wdduncan
Yes, you right @majensen!
I didn't mean to imply it was a jabberwocky thing. Only that that the installation directions might need to be updated with info about using venv in case someone tries to install in a virtual environment, gets an error and quits w/o further investigating the error.
