### 2020-05-18 #2168 @sap218
@wdduncan I fixed that `arise` issue so you should be able to run that command now :-) I will be improving the documentation today
