### 2020-05-18 702aa0c
adding better documentation to explain outputs
