+++ SCENARIO.md
@@ -1,0 +1,89 @@
+# example use-case of Jabberwocky
+
+#### **note** we are running these command in the `jabberwocky-test/process` directory...
+
+### scenario
+you have some textual data: **blog** posts from a social media platform. these posts include individuals discussing a topic which you are interested in. in this scenario the users are talking about [*pokemon*](https://simple.wikipedia.org/wiki/Pok%C3%A9mon).
+
+for example, you may have some posts (formatted differently) such as:
+> I think only gen 6 pokemon are on this path, try route 2 - wanderer wendy
+> No thanks, I'm, trying to catch a flying type in the mountains with the clear air - trainer penelope
+
+your **aim** is to extract particular **posts** which individuals use specific terms, e.g. "gen" or "flying".
+
+This is where **ontologies** are **useful**. ontologies are a controlled set of vocabulary with terms logically related to the other, e.g. our anatomy is that the iris is a part of the eye, but ontologies will have this defined in much more detail and logically. **however** the purpose of **jabberwocky** is not the logical relationships: but the **terms** themselves. ontologies store synonyms: an alternative version of a term, e.g. "eye" has synonyms, "eyeball", "globe", or "organ of sight". 
+
+in [jabberwocky-tests/ontology/](https://github.com/sap218/jabberwocky-tests/tree/master/ontology), there is the file, `pocketmonsters.owl`. here is the example ontology for our scenario, it can be expanded with additional terms or relationships however it is a useful start for the scenario.
+
+
+### catch
+using the `catch` command, you are trying to extract the posts which include key terms, e.g. "generation one", "dragon", and more. in [jabberwocky-tests/process/](https://github.com/sap218/jabberwocky-tests/tree/master/process) you will see the `listofwords.txt` which is a simple text file of these terms you are looking for, they match the exact terms from the ontology. there is also the textual data file: `public_forum` with a **total of 24 posts**, but split into 5 threads. 
+
+```
+$ catch -o ../ontology/pocketmonsters.owl -k listofwords.txt -t public_forum.json -p post > catch_01_output.txt
+```
+after using `catch` you will be given a file: `ontology_dict_class_synonyms.json` which is for your reference to observe the terms you want to use and their synonyms (for the scenario it was renamed to `catch_01_ontology_dict_class_synonyms.json`). it is worth noting that if you don't provide any keywords, the script will use every term from the ontology.
+
+**finally**, the `catch_01_output.txt` file a plain file with each post that includes one of the terms (or synonyms). this file will include **11 posts (out of 24)** which have the terms. 
+
+#### however
+you want to find out what terms are use most frequently, **plus** you notice in `catch_01_ontology_dict_class_synonyms.json` there lacks synonyms...
+
+
+### bite
+the `bite` command runs a tf-idf statistical analysis: searching for **important** (most frequrntly used) terms in a text corpus.
+```
+$ bite -t public_forum.json -p post
+```
+
+the main output is in [jabberwocky-tests/process/](https://github.com/sap218/jabberwocky-tests/tree/master/process), the file: `tfidf_results.csv` (renamed to `bite_01_tfidf_results.csv` for the scenario). this shows the tf-idf rankings in tabular form. this file is important to see which terms you may want to be included in your study, or to be added to the ontology as additional synonyms...
+
+| words   | count              |
+|---------|--------------------|
+| route   | 0.9693123548553331 |
+| pokemon | 0.8661273891167867 |
+| pokedex | 0.7561604589465487 |
+| path    | 0.7399254994619547 |
+| gen     | 0.7107711593211153 |
+| evolve  | 0.672293913917355  |
+
+
+### arise
[…]
