### 2020-05-18 4bb6205
updated README
the --user virtual environment fix
