### 2020-05-18 #2168 @sap218
@wdduncan i have also improved the documentation to include explanations of the outputs, see the [SCENARIO.md](https://github.com/sap218/jabberwocky-tests/blob/master/SCENARIO.md)
