### 2020-05-29 sap218/jabberwocky#12 @sap218
Will make a list of types of ontologies it works with!

~~Also the reason your own created ontology doesn't work is because your "hasExactSynonym" is not following the rdfs schema, it should be defined as "rdfs:hasExactSynonym" - I can include this is the documentation~~
