### 2020-05-29 sap218/jabberwocky#12 @sap218
I have included supported ontology formats in the README: `owl` including owl/xml and rdf/xml - I hope this helps @balhoff :-)

~~Edit: I also included information about RDF schema in the README~~
