### 2020-06-13 #2168 @wdduncan
I did another run through the documentation on github. It is looking much better! I really like the figure at the bottom :)

However, I still have issues/comments:
* In the help section on https://github.com/sap218/jabberwocky there is a typo in:
`-t, --textfile TEXT   **JSON ot TXT** file of text you want annotate. [required]` --> JSON **or** TXT file

* In the [scenario](https://github.com/sap218/jabberwocky-tests/blob/master/SCENARIO.md)
```
catch -o ../ontology/pocketmonsters.owl -k tests/catch/listofwords.txt -t public_forum.json -p post > catch_01_output.txt
```
`public_forum.json` is not in the catch directory; I modified the command to point to bite/public_forum.js

* In the command:
```
bite -t public_forum.json -p post
```
What does the `-p post` do? I couldn't find documentation about it. Sorry, if I missed it.

* When creating the table for `arise`, is there an ontology or controlled vocabulary that is supposed to define the values in the "type" column?

I looked at the documentation for [rdfs:schema](https://www.w3.org/TR/rdf-schema/)
But I did not find the annotation properties: rdfs:hasBroadSynonym, rdfs:hasNarrowSynonym, etc.

I also looked at the base IRI for [rdf-schema:hasBroadSynonym](https://www.w3.org/2000/01/rdf-schema#hasBroadSynonym) (i.e., https://www.w3.org/2000/01/rdf-schema) and still not find the aforementioned annotation properties.

Have you considered using [SKOS](https://www.w3.org/2004/02/skos/) which already has terms for broad match, close match, etc.
