### 2020-06-14 b625950
synonym appending confusion fixed
