### 2020-06-14 #2168 @sap218
@wdduncan I looked over your comments:

* I fixed that typo "ot".. now fixed to "or"
* You mentioned that when looking at scenario.md that public_forum.json is not in the catch directory, but the scenario states at the top, "we are running these command in the jabberwocky-test/process directory"
* I mention in the documentation that "-p, --parameter TEXT  parameter for the JSON file text."
* the 'type' column is a person's own defined synonyms - I have just fixed this confusion
* I will look into the rdfs:synonyms annotations

I hope this helps!
