### 2020-06-14 #2168 @wdduncan
@sap218
> "we are running these command in the jabberwocky-test/process directory".

Sorry. I missed this :)

> "-p, --parameter TEXT parameter for the JSON file text."

Yes, I saw that. Sorry, I wasn't clear. What does the `post` mean? Is it different that `post_text`? How is it different than `-p user-text`? E.g.: `bite --textfile social_media_posts.json --parameter user-text`.  Is there a controlled set of terms you use with `-p` flag?
