### 2020-06-14 #2168 @sap218
@wdduncan

`-p` refers to the field inside the JSON file for the text data, "-p post_text" and "-p user-text" are just two different examples - sorry about this confusion, i will change the README so all examples only refer to one example.
