### 2020-06-14 590041b
updated examples to fit the scenario.md
