### 2020-06-14 sap218/jabberwocky#12 @sap218
Update: turns out I was wrong about "rdfs:synonym" annotations - my apologies! Your example is correct, I just need to fix it :-)
I will look into making it user specified. Thanks!
