### 2020-06-14 #2168 @sap218
@wdduncan I have just fixed the annotation issue regarding the synonyms, I fixed the documentation to explain this:

> make sure annotations are defined with the oboInOWL: schema, e.g. hasExactSynonym should have the IRI http://www.geneontology.org/formats/oboInOWL#hasExactSynonym.
