### 2020-06-14 d64128a
Minor fix to arise (oboInOWL)
