### 2020-06-14 #2168 @wdduncan
@sap218 That's helpful. I think letting the users define custom relations is fine too. You just have to be clear about what is going on.

Here's a mapping project you might be interested: https://github.com/OBOFoundry/SSSOM
