### 2020-06-14 sap218/jabberwocky#12 @sap218
@balhoff I have now fixed the annotations issue, your ontology will work now if you include `<Prefix name="oboInOWL" IRI="http://www.geneontology.org/formats/oboInOWL#"/>` and ensure the `Owl` is all caps.
