### 2020-06-16 #2168 @majensen
> @wdduncan i have also improved the documentation to include explanations of the outputs, see the [SCENARIO.md](https://github.com/sap218/jabberwocky-tests/blob/master/SCENARIO.md)

(This is super cool.)
