### 2020-06-24 #2168 @wdduncan
@majensen
I checked boxes for installation and functionality. There aren't automated tests, but there is a tutorial. Are automated tests required?

@sap218
A number of changes have been made to the github pages. Does the manuscript reflect these changes? I haven't had a chance to re-read it yet.
