### 2020-06-24 #2168 @majensen
@wdduncan thanks - The tutorial can be used to verify functionality manually, so I think the letter of the law is satisfied. That said, @sap218 would you be willing to consider adding a few tests using [unittest](https://docs.python.org/3/library/unittest.html) or [pytest](https://docs.pytest.org/en/latest/index.html)? It is an important skill to have, if you haven't done it before, and it makes future support so much easier.
