### 2020-06-24 #2168 @sap218
@wdduncan The manuscript still reflects the software

@wdduncan @majensen

thank you all for the comments throughout and continued patience. I do understand the value of unit testing and plan to explore this aspect of software development when improving the package in the near future, but I am currently constrained by time because of my continued PhD work, which makes it difficult for me to develop this testing skill and update the package - but I hope everyone would agree that the testing repository (and scenario.md) provided does allow users to check the package works as intended (as we have done so for this review) and has sufficient steps describing the functionality of the software.

If the reviewers believe that adding unit tests on top of the system test provided by the test repository should be a pre-requisite of approving the manuscript then I will do my best to invest the time.
