### 2020-06-24 #2168 @balhoff
@majensen @sap218 I agree that the test/example repository is a fine functionality test; so I checked the tests box. I checked the last box, so I approve. Thank you both for your patience!
