+++ tox.ini
@@ -1,0 +1,10 @@
+[tox]
+envlist = py3
+
+[testenv]
+deps =
+     pytest
+     -r requirements.txt
+commands =
+      pytest
+
