### 2020-06-25 #2168 @majensen
Hey @sap218 - I added tests to jabberwocky; see [PR#13](https://github.com/sap218/jabberwocky/pull/13). These basically run the steps in the SCENARIO automatically, using the files in jabberwocky-tests. It's pretty basic, but now you have a working skeleton for more tests in the future. To run the tests, `pip install pytest` and then run `pytest` in the repo dir.
