### 2020-06-26 23578fe
updated README for contributor thanking
