### 2020-06-26 #2168 @sap218
@majensen Thank you so much! I just merged the PR - I appreciate the help! :-)

@balhoff Thank you!

@wdduncan We now have automated testing :-)
