### 2020-06-26 8a31597
included instructions for automated tests
