### 2020-06-29 #2168 @majensen
@wdduncan there are two boxes unchecked on your review . I think we've taken care of "Automated Tests". The last one is "State of the Field". If you're happy with where we are on these, can you confirm by checking those off. Then I can push our recommendation to the editors-in-chief.
