### 2020-06-29 #2168 @majensen
@sap218 we're on the final approach. Can you create a tag for your master branch, and then use that to create a public archive for your software? You can use [Zenodo](https://zenodo.org/) -- there's a quick tutorial at https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729
That process will give you a DOI (like "10.5281/zenodo.3558815" ) - please report that back here. Thanks!
