### 2020-06-29 #2168 @majensen
Excellent @sap218. Now, JOSS prefers that the name of the archive be the same as that of the paper, and that the author lists match. So.. would you be able to update the name and remove that parasitic author from the list on the archive?
