### 2020-06-30 #2168 @majensen
Hi @sap218 - can you check on the reference that whedon found missing in https://github.com/openjournals/joss-reviews/issues/2168#issuecomment-651877564 ?
