+++ paper.bib
@@ -132,5 +132,6 @@
   volume      =  445,
   pages       = "51--56",
   institution = "Austin, TX",
-  year        =  2010
+  year        =  2010,
+  doi	      = {10.25080/majora-92bf1922-00a}
 }
