### 2020-06-30 #2168 @majensen
sure @danielskatz , although you might tell @whedon (or @arfon) not to summon the eic until it performs a dry run free of errors. Sorry to disturb.
