### 2020-06-30 #2168 @danielskatz
👋 @sap218 - I have some comments/requests on the paper, in order from start to end:
- "gaining the key textual elements" - I don't know what this means; how does one gain textual elements?  Could "gaining" be "determining"?
- "having knowledge of the synonyms for clinical terminology may increase the size of the corpus available for analysis" - how is this the case?  How does having synonyms increase the size of the corpus?
- "e.g. social" <- "e.g., social"
- "their chosen ontology and" <- "their chosen ontology, and"
- "(Richardson, 2007)) and then" <- "(Richardson, 2007)), and then"
- "(e.g. lines from a corpus)" <- "(e.g., lines from a corpus)"
- "rank the important terms: using the" <- "rank the important terms by using the"
- "command arise will" <- "command, arise, will"
- "In Jabberwocky’s test repository (see Jabberwocky repo for further instructions), I show examples of each command separately. In the ‘process’ directory, I combine all three commands to demonstrate an example workflow. With 24 blog posts, the first use of catch returned 11 posts with the provided keywords. With bite I reviewed the CSV of ranked terms and curated new synonyms, simply by adding the corresponding class label from the ontology. I then used arise to add the identified synonyms into the ontology." <- "Jabberwocky’s test repository (see Jabberwocky repo for further instructions) shows examples of each command separately. The ‘process’ directory shows an example that combines all three commands to demonstrate an example workflow. With 24 blog posts, the first use of catch returned 11 posts with the provided keywords. The example uses bite to review the CSV of ranked terms and curated new synonyms, simply by adding the corresponding class label from the ontology. It then uses arise to add the identified synonyms into the ontology."
- For the reference Honnibal & Montani, "To appear" is not a journal. Please change this to include the journal name.
- For Manning et al., please protect cases with {}s in the bibtex so that the proceedings title appears in Title Case, rather than Sentence case. Also protect the {B} in the title so that it appears as Bloom, not bloom. Also Spacy should be spaCy, I think.
- In McKinney et al., do the same. Also protect the {P} in the title so that it appears as Python, not python. Also remove pdfs.semanticscholar.org
- In Pedregosa et al., fix Python in the title from python
- For Richardson, please add a URL or venue

When these are done, regenerate the pdf with ```@whedon generate pdf``` and let me know.
