+++ paper.bib
@@ -13,18 +13,16 @@
   doi      = {10.1093/database/bau033}
 }
 
-@ARTICLE{Honnibal2017-dn,
-  title   = "spacy 2: Natural language understanding with bloom embeddings,
+@MISC{Honnibal2017-dn,
+  title   = "spa{c}y 2: Natural language understanding with bloom embeddings,
              convolutional neural networks and incremental parsing",
   author  = "Honnibal, Matthew and Montani, Ines",
-  journal = "To appear",
-  volume  =  7,
-  number  =  1,
+  url	  = {https://github.com/explosion/spaCy}
   year    =  2017
 }
 
 @INPROCEEDINGS{Manning2014-rt,
-  title     = "The Stanford {CoreNLP} natural language processing toolkit",
+  title     = "The {S}tanford {CoreNLP} natural language processing toolkit",
   booktitle = "Proceedings of 52nd annual meeting of the association for
                computational linguistics: system demonstrations",
   author    = "Manning, Christopher and Surdeanu, Mihai and Bauer, John and
@@ -99,10 +97,11 @@
   journal   = "April",
   publisher = "media.readthedocs.org",
   year      =  2007
+  url	    = {https://beautiful-soup-4.readthedocs.io/en/latest/}
 }
 
 @ARTICLE{Pedregosa2011-st,
-  title    = "Scikit-learn: Machine Learning in Python",
+  title    = "Scikit-learn: Machine Learning in {P}ython",
   author   = "Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre
               and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and
               Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and
@@ -117,8 +116,8 @@
 }
 
 @INPROCEEDINGS{McKinney2010-xf,
-  title       = "Data structures for statistical computing in python",
-  booktitle   = "Proceedings of the 9th Python in Science Conference",
+  title       = "Data structures for statistical computing in {P}ython",
+  booktitle   = "Proceedings of the 9th {P}ython in {S}cience {C}onference",
   author      = "McKinney, Wes and {Others}",
   abstract    = "In this paper we are concerned with the practical issues of
                  working with data sets common to finance, statistics, and
@@ -128,7 +127,6 @@
                  models. We will discuss specific design issues encountered in
                  the course of developing pandas with relevant examples and
                  some comparisons with the R language.",
-  publisher   = "pdfs.semanticscholar.org",
   volume      =  445,
   pages       = "51--56",
   institution = "Austin, TX",
