+++ paper.md
@@ -22,15 +22,15 @@ bibliography: paper.bib
 
 # Summary
 
-Unstructured textual data is underused, as gaining the key textual elements is complicated by a lack of structured terms. To extract valuable parts from a corpus, a user will need to use a combination of terms and their associated synonyms. For example, when analysing clinical documents, having knowledge of the synonyms for clinical terminology may increase the size of the corpus available for analysis. Additionally, the type of synonyms needed depends on the text, e.g. social media posts will need awareness of layman’s terms. One way to represent our knowledge of terms associated with a domain is to create an ontology. Ontologies allow us to formalise our knowledge of a domain in a condensed manner by using controlled terms, called classes [@Hoehndorf2015-qr]. Classes can be annotated with metadata, including synonyms. Ontologies can include relationships between terms, and annotations such as cross-references to other ontologies [@Hoehndorf2015-qr].
+Unstructured textual data is underused, as gaining the key textual elements is complicated by a lack of structured terms, e.g., collecting the sentences from a corpus that are using particular terms. To extract valuable parts from a corpus, a user will need to use a combination of terms and their associated synonyms. For example, when analysing clinical documents, having knowledge of the synonyms for clinical terminology may increase the sentence output from the corpus, making more data available for analysis. Additionally, the type of synonyms needed depends on the text, e.g., social media posts will need awareness of layman’s terms. One way to represent our knowledge of terms associated with a domain is to create an ontology. Ontologies allow us to formalise our knowledge of a domain in a condensed manner by using controlled terms, called classes [@Hoehndorf2015-qr]. Classes can be annotated with metadata, including synonyms. Ontologies can include relationships between terms, and annotations such as cross-references to other ontologies [@Hoehndorf2015-qr].
 
 Clearly, ontologies are valuable for the analysis of textual data. Unfortunately, despite the existence of many well-established ontologies, such as the "Human Phenotype Ontology" [@Robinson2008-jh] and the "Disease Ontology" [@Schriml2012-qp], there remains a lack of tools that can take advantage of ontologies, especially for general text manipulation. Existing tools for annotating text, such as “spaCy” [@Honnibal2017-dn], “tagtog” [@Cejuela2014-lv], and “Stanford CoreNLP” [@Manning2014-rt] cannot interrogate text with an ontology directly, and require ontologies to be pre-processed into other formats (leaving the time-consuming task of extracting labels and tags from an ontology into a suitable intermediate format as an exercise for the end-user). These are specialist tools, returning all text in the document with every word tagged, as “noun”, “verb”, and other customised tags. There exists a niche for users who want to leverage an ontology to retrieve textual data from a corpus without having to perform any pre-processing, or parse away unwanted tags.
 
-We introduce Jabberwocky, a Python-based [@Van_Rossum1995-ia], open-source toolkit (accessible via https://github.com/sap218/jabberwocky) that allows users to query text in an ontology-aware fashion, and to modify those ontologies based on their findings. For example, with Jabberwocky’s ``catch`` command, a user provides textual data, their chosen ontology and a set of classes from the ontology to use as search terms. Jabberwocky cleans the input text, collects the annotated synonyms for the user-specified target classes (using “Beautiful Soup” to read the ontology’s XML structure [@Richardson2007-ba]) and then returns the key elements (e.g. lines from a corpus) which match one of the target terms, or a synonym from the ontology. The ``catch`` command will help users retrieve more matches for their chosen terms from the corpus, without users having to explicitly define all the possible synonyms or alternative spellings beforehand.
[…]
