### 2020-07-01 e350d56
fixing minor typos
