+++ paper.bib
@@ -17,7 +17,7 @@
   title   = "spa{c}y 2: Natural language understanding with bloom embeddings,
              convolutional neural networks and incremental parsing",
   author  = "Honnibal, Matthew and Montani, Ines",
-  url	  = {https://github.com/explosion/spaCy}
+  url	  = {https://github.com/explosion/spaCy},
   year    =  2017
 }
 
@@ -96,7 +96,7 @@
                to BS4.",
   journal   = "April",
   publisher = "media.readthedocs.org",
-  year      =  2007
+  year      =  2007,
   url	    = {https://beautiful-soup-4.readthedocs.io/en/latest/}
 }
 
