### 2020-07-01 3af7f06
minor fixes to bib (capitalisation)
