+++ paper.bib
@@ -14,7 +14,7 @@
 }
 
 @MISC{Honnibal2017-dn,
-  title   = "spa{c}y 2: Natural language understanding with bloom embeddings,
+  title   = "{s}pa{C}y 2: Natural language understanding with bloom embeddings,
              convolutional neural networks and incremental parsing",
   author  = "Honnibal, Matthew and Montani, Ines",
   url	  = {https://github.com/explosion/spaCy},
@@ -88,12 +88,6 @@
 @ARTICLE{Richardson2007-ba,
   title     = "Beautiful soup documentation",
   author    = "Richardson, Leonard",
-  abstract  = "You might be looking for the documentation for Beautiful Soup 3.
-               If so, you should know that Beautiful Soup 3 is no longer being
-               developed and that support for it will be dropped on or after
-               December 31, 2020. If you want to learn about the differences
-               between Beautiful Soup 3 and Beautiful Soup 4, see Porting code
-               to BS4.",
   journal   = "April",
   publisher = "media.readthedocs.org",
   year      =  2007,
@@ -119,14 +113,6 @@
   title       = "Data structures for statistical computing in {P}ython",
   booktitle   = "Proceedings of the 9th {P}ython in {S}cience {C}onference",
   author      = "McKinney, Wes and {Others}",
-  abstract    = "In this paper we are concerned with the practical issues of
-                 working with data sets common to finance, statistics, and
-                 other related fields. pandas is a new library which aims to
-                 facilitate working with these data sets and to provide a set
-                 of fundamental building blocks for implementing statistical
-                 models. We will discuss specific design issues encountered in
-                 the course of developing pandas with relevant examples and
-                 some comparisons with the R language.",
   volume      =  445,
   pages       = "51--56",
   institution = "Austin, TX",
