+++ paper.bib
@@ -14,7 +14,7 @@
 }
 
 @MISC{Honnibal2017-dn,
-  title   = "{s}pa{C}y 2: Natural language understanding with bloom embeddings,
+  title   = "{s}pa{C}y 2: Natural language understanding with {B}loom embeddings,
              convolutional neural networks and incremental parsing",
   author  = "Honnibal, Matthew and Montani, Ines",
   url	  = {https://github.com/explosion/spaCy},
