### 2020-07-01 bceb12c
redone the introduction
