### 2020-07-01 #2168 @sap218
@danielskatz I have updated the paper and addressed your comments
