### 2020-07-01 #2168 @danielskatz
Thanks -  I think one was missed, however: the proceedings title for Manning et al. should be in Title Case, not Sentence case.
