+++ paper.bib
@@ -23,8 +23,8 @@
 
 @INPROCEEDINGS{Manning2014-rt,
   title     = "The {S}tanford {CoreNLP} natural language processing toolkit",
-  booktitle = "Proceedings of 52nd annual meeting of the association for
-               computational linguistics: system demonstrations",
+  booktitle = "{P}roceedings of 52nd {A}nnual {M}eeting of the {A}ssociation for
+               {C}omputational {L}inguistics: {S}ystem {D}emonstrations",
   author    = "Manning, Christopher and Surdeanu, Mihai and Bauer, John and
                Finkel, Jenny and Bethard, Steven and McClosky, David",
   pages     = "55--60",
